# src/main/resources/static/docs/openapi3.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Orbit API",
    "description" : "오르비 백엔드 API 명세서입니다.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://dev.orbitalarm.net",
    "description" : "Dev Server"
  }, {
    "url" : "http://localhost:8080",
    "description" : "Local Server"
  } ],
  "tags" : [ ],
  "paths" : {
    "/api/v1/fortunes" : {
      "post" : {
        "tags" : [ "Fortune" ],
        "summary" : "운세 생성 API",
        "description" : "운세 생성 API",
        "operationId" : "userId가 숫자가 아닐 경우, 400 BAD REQUESTuserId가 유효하지 않을 경우, 400 BAD REQUEST외부 API 요청 중 오류 발생 시, 502 BAD GATEWAY운세 생성 200 OK운세 파싱 중 오류 발생 시, 500 INTERNAL SERVER ERROR존재하지 않는 사용자일 경우, 404 NOT FOUND",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "description" : "사용자 ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "404",
            "content" : {
              "text/plain;charset=UTF-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Exception"
                },
                "examples" : {
                  "존재하지 않는 사용자일 경우, 404 NOT FOUND" : {
                    "value" : "사용자를 찾을 수 없습니다."
                  }
                }
              }
            }
          },
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fortune"
                },
                "examples" : {
                  "운세 생성 200 OK" : {
                    "value" : "{\n  \"id\" : 1,\n  \"dailyFortune\" : \"Today is a great day!\",\n  \"avgFortuneScore\" : 90,\n  \"studyCareerFortune\" : {\n    \"score\" : 90,\n    \"title\" : \"title1\",\n    \"description\" : \"Success in studies\"\n  },\n  \"wealthFortune\" : {\n    \"score\" : 90,\n    \"title\" : \"title2\",\n    \"description\" : \"Prosperity in finance\"\n  },\n  \"healthFortune\" : {\n    \"score\" : 90,\n    \"title\" : \"title3\",\n    \"description\" : \"Strong physical health\"\n  },\n  \"loveFortune\" : {\n    \"score\" : 90,\n    \"title\" : \"title4\",\n    \"description\" : \"Harmonious love life\"\n  },\n  \"luckyOutfitTop\" : \"Red Shirt\",\n  \"luckyOutfitBottom\" : \"Black Pants\",\n  \"luckyOutfitShoes\" : \"White Sneakers\",\n  \"luckyOutfitAccessory\" : \"Gold Necklace\",\n  \"unluckyColor\" : \"Green\",\n  \"luckyColor\" : \"Blue\",\n  \"luckyFood\" : \"Pizza\"\n}"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "400",
            "content" : {
              "text/plain;charset=UTF-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Exception"
                },
                "examples" : {
                  "userId가 유효하지 않을 경우, 400 BAD REQUEST" : {
                    "value" : "user id는 null 또는 빈 문자열일 수 없습니다."
                  },
                  "userId가 숫자가 아닐 경우, 400 BAD REQUEST" : {
                    "value" : "user id는 숫자여야 합니다."
                  }
                }
              }
            }
          },
          "502" : {
            "description" : "502",
            "content" : {
              "text/plain;charset=UTF-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Exception"
                },
                "examples" : {
                  "외부 API 요청 중 오류 발생 시, 502 BAD GATEWAY" : {
                    "value" : "운세 데이터 요청 중 오류가 발생했습니다."
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "500",
            "content" : {
              "text/plain;charset=UTF-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Exception"
                },
                "examples" : {
                  "운세 파싱 중 오류 발생 시, 500 INTERNAL SERVER ERROR" : {
                    "value" : "운세 데이터를 처리하는 과정에서 오류가 발생했습니다."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prereservations" : {
      "post" : {
        "tags" : [ "PreReservation" ],
        "summary" : "사전 예약 생성 API",
        "description" : "사전 예약을 신청하는 API",
        "operationId" : "prereservations-create",
        "requestBody" : {
          "content" : {
            "application/json;charset=UTF-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/PreReservationCreateRequest"
              },
              "examples" : {
                "prereservations-create" : {
                  "value" : "{\n  \"email\" : \"byungwook-min@naver.com\",\n  \"phoneNumber\" : \"010-1234-5678\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "200"
          }
        }
      }
    },
    "/api/v1/users" : {
      "post" : {
        "tags" : [ "User" ],
        "summary" : "사용자 등록 API",
        "description" : "사용자를 등록하는 API",
        "operationId" : "users-create",
        "requestBody" : {
          "content" : {
            "application/json;charset=UTF-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/SaveUserRequest"
              },
              "examples" : {
                "users-create" : {
                  "value" : "{\n  \"name\" : \"홍길동\",\n  \"birthDate\" : \"2025-02-09\",\n  \"birthTime\" : \"08:30:00\",\n  \"calendarType\" : \"SOLAR\",\n  \"gender\" : \"MALE\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserId"
                },
                "examples" : {
                  "users-create" : {
                    "value" : "1"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fortunes/{fortuneId}" : {
      "get" : {
        "tags" : [ "Fortune" ],
        "summary" : "운세 조회 API",
        "description" : "운세 조회 API",
        "operationId" : "운세 조회 200 OK존재하지 않는 운세인 경우, 404 NOT FOUND",
        "parameters" : [ {
          "name" : "fortuneId",
          "in" : "path",
          "description" : "오늘 생성된 운세 ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "404",
            "content" : {
              "text/plain;charset=UTF-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Exception"
                },
                "examples" : {
                  "존재하지 않는 운세인 경우, 404 NOT FOUND" : {
                    "value" : "해당 운세를 찾을 수 없습니다. 운세 생성을 요청해주세요."
                  }
                }
              }
            }
          },
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fortune"
                },
                "examples" : {
                  "운세 조회 200 OK" : {
                    "value" : "{\n  \"id\" : 1,\n  \"dailyFortune\" : \"Today is a great day!\",\n  \"avgFortuneScore\" : 90,\n  \"studyCareerFortune\" : {\n    \"score\" : 90,\n    \"title\" : \"title1\",\n    \"description\" : \"Success in studies\"\n  },\n  \"wealthFortune\" : {\n    \"score\" : 90,\n    \"title\" : \"title2\",\n    \"description\" : \"Prosperity in finance\"\n  },\n  \"healthFortune\" : {\n    \"score\" : 90,\n    \"title\" : \"title3\",\n    \"description\" : \"Strong physical health\"\n  },\n  \"loveFortune\" : {\n    \"score\" : 90,\n    \"title\" : \"title4\",\n    \"description\" : \"Harmonious love life\"\n  },\n  \"luckyOutfitTop\" : \"Red Shirt\",\n  \"luckyOutfitBottom\" : \"Black Pants\",\n  \"luckyOutfitShoes\" : \"White Sneakers\",\n  \"luckyOutfitAccessory\" : \"Gold Necklace\",\n  \"unluckyColor\" : \"Green\",\n  \"luckyColor\" : \"Blue\",\n  \"luckyFood\" : \"Pizza\"\n}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "사용자 조회 API",
        "description" : "사용자 정보를 조회하는 API",
        "operationId" : "users-get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "사용자 ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                },
                "examples" : {
                  "users-get" : {
                    "value" : "{\n  \"id\" : 1,\n  \"name\" : \"홍길동\",\n  \"birthDate\" : \"2025-02-09\",\n  \"birthTime\" : \"08:30\",\n  \"calendarType\" : \"SOLAR\",\n  \"gender\" : \"MALE\"\n}"
                  }
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "User" ],
        "summary" : "사용자 수정 API",
        "description" : "사용자 정보를 수정하는 API (전체 덮어쓰기 방식)",
        "operationId" : "users-update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "수정할 사용자 ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json;charset=UTF-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateUserRequest"
              },
              "examples" : {
                "users-update" : {
                  "value" : "{\n  \"name\" : \"홍길동\",\n  \"birthDate\" : \"2025-02-09\",\n  \"birthTime\" : \"08:30:00\",\n  \"calendarType\" : \"SOLAR\",\n  \"gender\" : \"MALE\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "description" : "204"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "UpdateUserRequest" : {
        "title" : "UpdateUserRequest",
        "required" : [ "birthDate", "birthTime", "calendarType", "gender", "name" ],
        "type" : "object",
        "properties" : {
          "calendarType" : {
            "type" : "string",
            "description" : "달력 타입 (SOLAR, LUNAR 등)"
          },
          "gender" : {
            "type" : "string",
            "description" : "사용자의 성별 (MALE, FEMALE 등)"
          },
          "name" : {
            "type" : "string",
            "description" : "사용자의 이름"
          },
          "birthTime" : {
            "type" : "string",
            "description" : "사용자의 출생시간 (HH:mm)"
          },
          "birthDate" : {
            "type" : "string",
            "description" : "사용자의 생년월일 (yyyy-MM-dd)"
          }
        }
      },
      "Exception" : {
        "title" : "Exception",
        "type" : "object"
      },
      "SaveUserRequest" : {
        "title" : "SaveUserRequest",
        "required" : [ "birthDate", "birthTime", "calendarType", "gender", "name" ],
        "type" : "object",
        "properties" : {
          "calendarType" : {
            "type" : "string",
            "description" : "달력 타입 (SOLAR, LUNAR 등)"
          },
          "gender" : {
            "type" : "string",
            "description" : "사용자의 성별 (MALE, FEMALE 등)"
          },
          "name" : {
            "type" : "string",
            "description" : "사용자의 이름"
          },
          "birthTime" : {
            "type" : "string",
            "description" : "사용자의 출생시간 (HH:mm)"
          },
          "birthDate" : {
            "type" : "string",
            "description" : "사용자의 생년월일 (yyyy-MM-dd)"
          }
        }
      },
      "UserId" : {
        "title" : "UserId",
        "type" : "object"
      },
      "User" : {
        "title" : "User",
        "required" : [ "birthDate", "birthTime", "calendarType", "gender", "id", "name" ],
        "type" : "object",
        "properties" : {
          "calendarType" : {
            "type" : "string",
            "description" : "달력 타입 (SOLAR, LUNAR 등)"
          },
          "gender" : {
            "type" : "string",
            "description" : "사용자의 성"
          },
          "name" : {
            "type" : "string",
            "description" : "사용자의 이름"
          },
          "birthTime" : {
            "type" : "string",
            "description" : "사용자의 출생시간 (HH:mm)"
          },
          "id" : {
            "type" : "number",
            "description" : "사용자 ID"
          },
          "birthDate" : {
            "type" : "string",
            "description" : "사용자의 생년월일 (yyyy-MM-dd)"
          }
        }
      },
      "PreReservationCreateRequest" : {
        "title" : "PreReservationCreateRequest",
        "required" : [ "email", "phoneNumber" ],
        "type" : "object",
        "properties" : {
          "phoneNumber" : {
            "type" : "string",
            "description" : "사전 예약을 신청하는 사용자의 전화번호"
          },
          "email" : {
            "type" : "string",
            "description" : "사전 예약을 신청하는 사용자의 이메일 주소"
          }
        }
      },
      "Fortune" : {
        "title" : "Fortune",
        "required" : [ "avgFortuneScore", "dailyFortune", "id", "luckyColor", "luckyFood", "luckyOutfitAccessory", "luckyOutfitBottom", "luckyOutfitShoes", "luckyOutfitTop", "unluckyColor" ],
        "type" : "object",
        "properties" : {
          "luckyOutfitBottom" : {
            "type" : "string",
            "description" : "행운의 의상 하의"
          },
          "luckyColor" : {
            "type" : "string",
            "description" : "행운의 색깔"
          },
          "wealthFortune" : {
            "required" : [ "description", "score", "title" ],
            "type" : "object",
            "properties" : {
              "score" : {
                "type" : "number",
                "description" : "재물운 점수"
              },
              "description" : {
                "type" : "string",
                "description" : "재물운 설명"
              },
              "title" : {
                "type" : "string",
                "description" : "재물운 제목"
              }
            }
          },
          "healthFortune" : {
            "required" : [ "description", "score", "title" ],
            "type" : "object",
            "properties" : {
              "score" : {
                "type" : "number",
                "description" : "건강운 점수"
              },
              "description" : {
                "type" : "string",
                "description" : "건강운 설명"
              },
              "title" : {
                "type" : "string",
                "description" : "건강운 제목"
              }
            }
          },
          "luckyOutfitShoes" : {
            "type" : "string",
            "description" : "행운의 의상 신발"
          },
          "unluckyColor" : {
            "type" : "string",
            "description" : "불운의 색깔"
          },
          "luckyOutfitTop" : {
            "type" : "string",
            "description" : "행운의 의상 상의"
          },
          "luckyFood" : {
            "type" : "string",
            "description" : "행운의 음식"
          },
          "luckyOutfitAccessory" : {
            "type" : "string",
            "description" : "행운의 의상 액세서리"
          },
          "avgFortuneScore" : {
            "type" : "number",
            "description" : "운세 평균 점수"
          },
          "studyCareerFortune" : {
            "required" : [ "description", "score", "title" ],
            "type" : "object",
            "properties" : {
              "score" : {
                "type" : "number",
                "description" : "학업운 점수"
              },
              "description" : {
                "type" : "string",
                "description" : "학업운 설명"
              },
              "title" : {
                "type" : "string",
                "description" : "학업운 제목"
              }
            }
          },
          "dailyFortune" : {
            "type" : "string",
            "description" : "오늘의 운세 메시지"
          },
          "id" : {
            "type" : "number",
            "description" : "운세 ID"
          },
          "loveFortune" : {
            "required" : [ "description", "score", "title" ],
            "type" : "object",
            "properties" : {
              "score" : {
                "type" : "number",
                "description" : "사랑운 점수"
              },
              "description" : {
                "type" : "string",
                "description" : "사랑운 설명"
              },
              "title" : {
                "type" : "string",
                "description" : "사랑운 제목"
              }
            }
          }
        }
      }
    }
  }
}